API/requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
numpy>=1.19.0
scipy>=1.5.0
Pillow>=8.0.0
torch>=1.7.0
torchvision>=0.8.0
werkzeug>=2.3.0


